test(BA-5943): consolidate model card delete tests under the repository layer
Drop the adapter-level mock test and extend the real-DB
``test_model_card_delete`` with both option_off and option_on cases. The
adapter only forwards the option to the action, so behavioral coverage
naturally belongs at the repository layer where the cascade actually
happens.
